Tiger implementation C++ => C99
I realized the only thing keeping Tiger a C++ implementation was one
single pass-by-reference function.  I'm biting the bullet and switching
it to a macro.

git-svn-id: https://fluks.googlecode.com/svn@190 6b01645e-0d02-11de-a42d-873b86d837ff

+++ twofish.c
@@ -55,7 +55,7 @@ const uint32_t RHO = 0x01010101UL;
 #define b2(x) (uint8_t)((x) >> 16)
 /* don't need to mask since all bits are in lower 8 - byte cast here saves
    nothing, but hey, what the hell, it doesn't hurt any */
-#define b3(x) (uint8_t)((x) >> 24)
+#define b3(x) ((x) >> 24)
 
 /*
    multiply two polynomials represented as uint32_t's, actually called with uint8_tS,
